Fixing Issue#480 (#496)
* Fixing Issue#480

* Elijah's Feedback

* JavaDoc description for skipping test

* Removed @Test annotation, waiting on #474 to be merged

* Final Changes

* Refactoring, Add @Test dependencies
